# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Saving Grains API
  description: |
    The Saving Grains API provides endpoints for user authentication, user management, and wallet request functionality. 

    ## Authentication
    Most endpoints require authentication via JWT token passed in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## Error Handling
    All API errors follow a consistent format with an `error` field containing the error message.
  version: 1.0.0
  contact:
    name: Saving Grains Support
    email: support@savinggrains.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.savinggrains.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          description: User's unique identifier
          example: "507f1f77bcf86cd799439011"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        phone_number:
          type: string
          example: "+233123456789"
        roles:
          type: array
          items:
            type: string
          example: ["user", "admin"]
        system:
          type: boolean
          example: "false"
        createdAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00.000Z"
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message description
          example: "Invalid credentials"
...
    CreateUserRequest:
      type: object
      required:
        - first_name
        - last_name
        - email
        - phone_number
        - password
      properties:
        first_name:
          type: string
          description: User's first name
          example: "John"
        last_name:
          type: string
          description: User's last name
          example: "Doe"
        email:
          type: string
          format: email
          description: User's email address (must be unique)
          example: "john.doe@example.com"
        phone_number:
          type: string
          description: User's phone number (must be unique)
          example: "+233123456789"
        password:
          type: string
          description: User's password (will be hashed)
          example: "securePassword123"
        roles:
          type: array
          items:
            type: string
          description: Optional array of user roles
          example: ["user"]
...
paths:
...
  /users:
    get:
      summary: Get all users
      description: Retrieves a list of all users in the system
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Failed to fetch users"

    post:
      summary: Create a new user
      description: Creates a new user account with the provided information
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Missing required fields"
        "409":
          description: Conflict - user already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "User already exists"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "Failed to create user"
...
security:
  - bearerAuth: []